requirements.txt
discord.py
python-dotenv
openai
coloredlogs
yt_dlp
PyNaCl